\name{SPA.example}
\alias{SPA.example}
\docType{data}
\title{ An example data set
}
\description{
The example data set containts fomatted phenotype and genotype data.

}
\usage{data(SPA.example)}
\format{
  The format is:
List of 2
 $ pheno: Named num [1:500] 1.74 1.25 1.53 1.94 1.73 ...
  ..- attr(*, "names")= chr [1:500] "1" "2" "3" "4" ...
 $ geno : int [1:500, 1:4] 1 1 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:500] "1" "2" "3" "4" ...
  .. ..$ : chr [1:4] "X634820282" "X634820324" "X632197333" "X632197358"
}

\examples{
data(SPA.example)
}
\keyword{datasets}
